se reestructura las funciones para que funcionen (valga la redundancia) de manera general en las diversas rutas existentes
